+++ src/01-document/isDisambig.js
@@ -6,6 +6,20 @@ const i18n_templates = i18n.disambig_templates.reduce((h, str) => {
   return h
 }, {})
 
+// look for '... may refer to'
+const byText = function (s) {
+  if (!s) {
+    return false
+  }
+  let txt = s.text()
+  if (txt !== null && txt[0]) {
+    if (/. may (also)? refer to\b/i.test(txt) === true) {
+      return true
+    }
+  }
+  return false
+}
+
 /**
  * Parses the wikitext to find out if this page is a disambiguation
  *
@@ -28,14 +42,8 @@ const isDisambig = function (doc) {
     return true
   }
   //try 'may refer to' on first line for en-wiki?
-  let s = doc.sentences()[0]
-  if (s) {
-    let firstLine = s.text()
-    if (firstLine !== null && firstLine[0]) {
-      if (/. may refer to\b/i.test(firstLine) === true) {
-        return true
-      }
-    }
+  if (byText(doc.sentence(0)) === true || byText(doc.sentence(1)) === true) {
+    return true
   }
   return false
 }